fix: preserve embed params on back-to-search navigation and consolidate embedded state functions
Problem: Embed params (uiEmbed=v0, uiSearchHideGraph=1, etc.) were silently dropped when navigating Search → Trace → Back to Search, breaking embedded UI context.

Solution:
1. Split single computed URL into two separate URLs in SearchResults:
   - backNavUrl: preserves embed params for 'Back to Search' navigation
   - standaloneUrl: strips embed params for standalone new-tab link
2. Moved getEmbeddedState() and stripEmbeddedState() from utils/embedded-url.ts to stores/embedded-store.ts (eliminating unnecessary indirection)
3. Added forbidNewPage guard to prevent rendering standalone link when forbidNewPage config is true
4. Updated tests to reflect new behavior and consolidate test coverage

This ensures users maintain embedded context when clicking 'Back to Search' while intentionally stripping embed context for standalone links.
